Add datanode heartbeating on write
