+++ crates/hdfs-native/src/hdfs/mod.rs
@@ -1,4 +1,5 @@
+pub(crate) mod block_reader;
+pub(crate) mod block_writer;
 pub(crate) mod connection;
-pub(crate) mod datanode;
 pub(crate) mod protocol;
 pub(crate) mod proxy;
